Added initial support for helpers. Hooray.
Helpers are executables that take function/method arguments and pass them as arguments to helper script. Keyword arguments are passed as env (TODO: VIRTUI_ prefix needs to be added). Communication with user may be done via stdin and stdout and stderr is meant as result of helper script for virtui.